completed refactoring of gulpfile
